+++ src/index.ts
@@ -3,6 +3,7 @@ export { default as RmConnection } from './rmConnection';
 export { default as RmPool } from './rmPool';
 export { default as RmPoolConnection } from './rmPoolConnection';
 export { RmPools } from './rmPools';
+export { RmLogger } from './logger';
 
 // Export types
 export * from './types';